chore: update stageToSecond method & counterView

+++ app/src/main/java/com/example/dgtimer/activities/timer/Counter.kt
@@ -2,8 +2,8 @@ package com.example.dgtimer.activities.timer
 
 data class Counter(
     val type: String,
-    val totalTime: Int,
+    val totalTime: Long,
     val index: Int,
     val isActive: Boolean,
-    val currentTime: Int = totalTime,
+    val currentTime: Long = totalTime,
 )
